pdf_extract_kit/tasks/table_parsing/models/paddle_table.py
# pdf_extract_kit/tasks/table_parsing/models/paddle_table.py

import os
from PIL import Image

class PaddleTableParsing:
    def __init__(self, config):
        self.config = config
        print("Initialized PaddleTableParsing with config:", config)

    def predict(self, images, result_path, output_format="latex", **kwargs):
        os.makedirs(result_path, exist_ok=True)

        results = {}
        for i, img_path in enumerate(images):
            print(f"Processing {img_path}...")

            # Load image
            image = Image.open(img_path)

            # Placeholder: generate dummy LaTeX
            latex_code = (
                "\\begin{tabular}{|c|c|}\n"
                "\\hline\n"
                "Header1 & Header2 \\\\\n"
                "\\hline\n"
                "Value1 & Value2 \\\\\n"
                "Value3 & Value4 \\\\\n"
                "\\hline\n"
                "\\end{tabular}"
            )

            result_file = os.path.join(result_path, f"table_{i:03}.tex")
            with open(result_file, "w", encoding="utf-8") as f:
                f.write(latex_code)

            results[img_path] = latex_code

        return results
