requirements.txt
omegaconf
matplotlib
PyMuPDF
ultralytics>=8.2.85
torch==2.2.2
doclayout-yolo==0.0.4
unimernet==0.2.1
paddlepaddle
paddleocr==2.7.3
struct-eqtable
Pillow
